stage2: remove needless cast
Fixes the following clippy warning:

warning: casting integer literal to `usize` is unnecessary
  --> src/stage2.rs:89:30
   |
89 |         VirtAddr::from(640 * 1024 as usize),
   |                              ^^^^^^^^^^^^^ help: try: `1024_usize`
   |
   = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
   = note: `#[warn(clippy::unnecessary_cast)]` on by default

+++ src/stage2.rs
@@ -86,7 +86,7 @@ fn setup_env() {
     install_console_logger("Stage2");
     init_kernel_mapping_info(
         VirtAddr::null(),
-        VirtAddr::from(640 * 1024 as usize),
+        VirtAddr::from(640 * 1024usize),
         PhysAddr::null(),
     );
     register_cpuid_table(unsafe { &CPUID_PAGE });
